refactor: enhance indexing status table and remove deprecated components
